requirements.txt
numpy==1.20.3
python-mnist